+++ lib/data-generators/file-hash.js
@@ -31,14 +31,6 @@ module.exports = CoreObject.extend({
       .then(function(contents) {
         return md5Hash(contents.toString());
       })
-  },
-
-  _resolveConfigValue: function(key, config, context) {
-    if(typeof config[key] === 'function') {
-      return config[key](context);
-    }
-
-    return config[key];
   }
 });
