+++ lib/data-generators/git-tag-commit.js
