+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ember-cli-deploy-revision-data",
   "version": "0.1.0-beta.2",
-  "description": "An ember-cli-plugin to create a unique revision key for the build",
+  "description": "An ember-cli-deploy plugin to generate data about this deploy revision including a unique revision key based on the current build",
   "directories": {
     "doc": "doc",
     "test": "tests"
