+++ tests/unit/lib/data-generators/git-tag-commit-nodetest.js
@@ -3,12 +3,12 @@
 var assert = require('ember-cli/tests/helpers/assert');
 var gitRepoInfo = require('git-repo-info');
 
-describe('the git-tag-commit key generator', function() {
-  var KeyGenerator;
+describe('the git-tag-commit data generator', function() {
+  var DataGenerator;
   var cwd;
 
   before(function() {
-    KeyGenerator = require('../../../../lib/key-generators/git-tag-commit');
+    DataGenerator = require('../../../../lib/data-generators/git-tag-commit');
     gitRepoInfo._changeGitDir('dotgit');
   });
 
@@ -24,7 +24,7 @@ describe('the git-tag-commit key generator', function() {
     it('concatenates the git tag and the git commit hash', function() {
       process.chdir('tests/fixtures/repo');
 
-      var subject = new KeyGenerator();
+      var subject = new DataGenerator();
 
       return assert.isFulfilled(subject.generate())
         .then(function(revision) {
@@ -35,7 +35,7 @@ describe('the git-tag-commit key generator', function() {
     it('rejects if no repository found', function() {
       process.chdir('tests/fixtures/not-a-repo');
 
-      var subject = new KeyGenerator();
+      var subject = new DataGenerator();
 
       return assert.isRejected(subject.generate())
         .then(function(error) {
@@ -46,7 +46,7 @@ describe('the git-tag-commit key generator', function() {
     it('rejects if no git tag found', function() {
       process.chdir('tests/fixtures/tagless-repo');
 
-      var subject = new KeyGenerator();
+      var subject = new DataGenerator();
 
       return assert.isRejected(subject.generate())
         .then(function(error) {
